docs(blog): add "WebGL Without a GPU" deep dive
Technical post on rendering WebGL screenshots on a GPU-less fleet:
ANGLE -> Mesa llvmpipe, the Xvfb surface requirement, the from-source
Mesa build, and the silent 2D fallback guarded in CI.

- Embed browserless.report() using the integrations/cli JSON visualizer,
  extracted into a shared components/elements/JsonView and reused on the
  CLI page with no visual change.
- Add components/markdown/BrowserlessReport with a real report payload.
- Capitalize GPU/Mesa in title casing.
