+++ data/git-timestamps-created.json
@@ -82,7 +82,7 @@
   "src/content/docs/api/parameters/prerender.md": "2026-01-06T00:00:00.000Z",
   "src/content/docs/api/parameters/proxy.md": "2026-01-06T00:00:00.000Z",
   "src/content/docs/api/parameters/retry.md": "2026-01-06T00:00:00.000Z",
-  "src/content/docs/api/parameters/screenshot/animated.md": "2026-06-28T00:00:00.000Z",
+  "src/content/docs/api/parameters/screenshot/animated.md": "2026-06-29T00:00:00.000Z",
   "src/content/docs/api/parameters/screenshot/codeScheme.md": "2026-01-06T00:00:00.000Z",
   "src/content/docs/api/parameters/screenshot/element.md": "2026-01-06T00:00:00.000Z",
   "src/content/docs/api/parameters/screenshot/fullPage.md": "2026-01-06T00:00:00.000Z",
