+++ data/git-timestamps-modified.json
@@ -86,7 +86,7 @@
   "src/content/docs/api/parameters/screenshot/codeScheme.md": "2026-03-17T00:00:00.000Z",
   "src/content/docs/api/parameters/screenshot/element.md": "2026-02-04T00:00:00.000Z",
   "src/content/docs/api/parameters/screenshot/fullPage.md": "2026-02-04T00:00:00.000Z",
-  "src/content/docs/api/parameters/screenshot/index.md": "2026-06-28T00:00:00.000Z",
+  "src/content/docs/api/parameters/screenshot/index.md": "2026-06-29T00:00:00.000Z",
   "src/content/docs/api/parameters/screenshot/omitBackground.md": "2026-02-04T00:00:00.000Z",
   "src/content/docs/api/parameters/screenshot/overlay.md": "2026-04-22T00:00:00.000Z",
   "src/content/docs/api/parameters/screenshot/type.md": "2026-02-04T00:00:00.000Z",
