+++ src/components/pages/cli/TerminalJsonView.js
@@ -1,7 +1,6 @@
-import React, { useEffect, useState } from 'react'
+import React from 'react'
 
-import Box from 'components/elements/Box'
-import { colors, theme } from 'theme'
+import JsonView from 'components/elements/JsonView/JsonView'
 
 const HN_PAGE_HEADERS = {
   server: 'nginx',
@@ -122,147 +121,6 @@ const CLI_RESPONSE = {
   }
 }
 
-const CLI_TERMINAL_JSON_THEME = {
-  base00: 'rgba(0, 0, 0, 0)',
-  base01: colors.white,
-  base02: colors.black10,
-  base03: colors.black50,
-  base04: colors.black50,
-  base05: colors.black80,
-  base06: colors.black,
-  base07: colors.black,
-  base08: colors.red6,
-  base09: colors.green6,
-  base0A: colors.black50,
-  base0B: colors.blue6,
-  base0C: colors.violet4,
-  base0D: colors.black80,
-  base0E: colors.green5,
-  base0F: colors.blue6
-}
-
-const JSON_DEFAULT_EXPAND_LEVEL = 2
-
-// react-json-view only applies shouldCollapse when collapsed > depth; +1 unlocks
-// per-field control at the deepest default level (e.g. response.body).
-const JSON_COLLAPSED_DEPTH = JSON_DEFAULT_EXPAND_LEVEL + 1
-
-const collapseField = field => {
-  const namespaceDepth = field.namespace?.length ?? 0
-
-  if (namespaceDepth <= JSON_DEFAULT_EXPAND_LEVEL) return false
-
-  if (field.name === 'body') return false
-
-  return (
-    field.type === 'object' ||
-    (field.type === 'array' && Array.isArray(field.src) && field.src.length > 0)
-  )
-}
-
-const jsonViewCss = theme({
-  fontFamily: 'mono',
-  fontSize: '13px',
-  lineHeight: '20px',
-  '& .react-json-view': {
-    backgroundColor: 'transparent !important',
-    fontFamily: 'inherit',
-    fontSize: 'inherit',
-    lineHeight: 'inherit',
-    padding: 0,
-    margin: 0
-  },
-  '& .react-json-view div > span': {
-    padding: 0
-  },
-  '& .pretty-json-container > .object-content > .object-key-val': {
-    paddingTop: 0,
-    paddingBottom: 0,
-    paddingLeft: '0 !important',
-    marginLeft: 0,
-    borderLeft: '0 !important'
-  },
-  '& .icon-container': {
-    display: 'inline-flex',
-    alignItems: 'center',
-    justifyContent: 'center',
-    verticalAlign: 'baseline',
-    position: 'relative',
-    top: '1px',
-    width: '14px',
-    height: '1em',
-    lineHeight: 1,
-    marginLeft: 0,
-    paddingLeft: 0,
-    flexShrink: 0
-  },
-  '& .expanded-icon, & .collapsed-icon': {
-    display: 'inline-flex',
-    alignItems: 'center',
-    lineHeight: 1
-  },
-  '& .expanded-icon svg, & .collapsed-icon svg': {
-    display: 'block',
-    verticalAlign: 'middle',
-    paddingLeft: '0 !important'
-  },
-  '& .object-key-val .object-key-val': {
-    paddingTop: 0,
-    paddingBottom: 0,
-    borderLeft: `1px solid ${colors.black10}`
-  },
-  '& .brace-row, & .icon-container': {
-    cursor: 'pointer'
-  }
-})
-
-const TerminalJsonView = () => {
-  const [ReactJson, setReactJson] = useState(null)
-
-  useEffect(() => {
-    let cancelled = false
-
-    import('@microlink/react-json-view').then(module => {
-      if (!cancelled) setReactJson(() => module.default)
-    })
-
-    return () => {
-      cancelled = true
-    }
-  }, [])
-
-  if (!ReactJson) return null
-
-  return (
-    <Box css={jsonViewCss}>
-      <ReactJson
-        src={CLI_RESPONSE}
-        name={false}
-        theme={CLI_TERMINAL_JSON_THEME}
-        collapsed={JSON_COLLAPSED_DEPTH}
-        shouldCollapse={collapseField}
-        collapseStringsAfterLength={72}
-        quotesOnKeys={false}
-        displayDataTypes={false}
-        displayObjectSize={false}
-        displayArrayKey={false}
-        enableClipboard={false}
-        onEdit={false}
-        onAdd={false}
-        onDelete={false}
-        onSelect={false}
-        sortKeys={false}
-        indentWidth={2}
-        iconStyle='triangle'
-        style={{
-          backgroundColor: 'transparent',
-          fontFamily: 'inherit',
-          fontSize: 'inherit',
-          lineHeight: 'inherit'
-        }}
-      />
-    </Box>
-  )
-}
+const TerminalJsonView = () => <JsonView src={CLI_RESPONSE} />
 
 export default TerminalJsonView
