+++ src/helpers/title.js
@@ -1,12 +1,8 @@
 import capitalize from 'microsoft-capitalize'
 
 const special = [
-  'ApiFlash',
-  'ScreenshotAPI',
-  'Screenshotmachine',
-  'ScreenshotOne',
-  'URL2PNG',
   'AI',
+  'ApiFlash',
   'apiKey',
   'APIs',
   'background',
@@ -20,6 +16,7 @@ const special = [
   'fullPage',
   'GA',
   'GitHub',
+  'GPU',
   'hasTouch',
   'height',
   'HTML',
@@ -28,6 +25,7 @@ const special = [
   'JavaScript',
   'Lighthouse Viewer',
   'Lighthouse',
+  'Mesa',
   'Microlink Cards',
   'Microlink',
   'oEmbed',
@@ -37,6 +35,9 @@ const special = [
   'PDFs',
   'quality',
   'retry',
+  'ScreenshotAPI',
+  'Screenshotmachine',
+  'ScreenshotOne',
   'ScreenshotOne',
   'scroll',
   'SDK',
@@ -47,6 +48,7 @@ const special = [
   'TypeScript',
   'Unified Cache',
   'url',
+  'URL2PNG',
   'URLs',
   'waitFor',
   'waitUntil',
